+++ ChangeLog.md
@@ -1,7 +1,8 @@
 ## 1.2.0 - 2018-03-
 
 - JL0006 and JL0007 fixed: empty HTML attributes correctly detected
-
+- added "-warning-to-error" to enforce all warnings
+- added "-error-to-warning" to support smooth migration of existing pages with findings
 
 ## 1.1.0 - 2018-03-03